move java check to server
